chore: split functions endpoint in digestible files
